Make sure block data is actually on disk
This mimics how S3 and the S3 module behaves: If a block was written
successfully we can be reasonably sure that it can be retrieved
correctly. Doing fdatasync() on every block will impact performance,
but without it a backup could complete successfully, even though not
all block data has reached stable storage.