Fix DetachedInstanceError when calling benji enforce with --machine-output
The problem is twofold: Due to the fact that SQLAlchemy expires objects
on commit reading the labels via the version object fails after the
version has been deleted in the database. This is fixed by
expire_on_commit=False. But then reading the labels itself fails because
they were never present in the session due to the default of lazily
loading relationships. This is fixed by making sure to always load the
associated labels when a version is loaded from the database.