Set a `grouping_key` when exporting prometheus metrics
Set a grouping_key when exporting prometheus metrics to ensure that metrics
for different runs are not overwritten.

Fix a bug in `backup_pvc` where metrics are not exported on ceph failure.
Instead, just push to the gatewhere right where the metrics are incremented.

Note: There is still an additional issue not handled here as a ceph failure
will raise an exception and skip backing up any additional PVCs in the
namespace, so it won't export any metrics for those.