Remove check of imported key's curve name against configured curve name
Looking at the code of PyCryptome it seems that the key import function matches
the curve's oid by traversing a dictionary of curves indexed by curve name. The
dictionary contains one entry for each alias name that a curve has. So
ecc_key.curve might return any one of these aliases. As the list of curves is
internal to PyCryptome we cannot look up a curve and all of its aliases.